+++ mockserver-core/src/test/java/org/mockserver/serialization/model/ConnectionOptionsDTOTest.java
@@ -2,7 +2,9 @@ package org.mockserver.serialization.model;
 
 import org.junit.Test;
 import org.mockserver.model.ConnectionOptions;
+import org.mockserver.model.Delay;
 
+import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
@@ -13,12 +15,13 @@ public class ConnectionOptionsDTOTest {
     public void shouldReturnValuesSetInConstructor() {
         // when
         ConnectionOptionsDTO connectionOptions = new ConnectionOptionsDTO(
-                new ConnectionOptions()
-                        .withSuppressContentLengthHeader(true)
-                        .withContentLengthHeaderOverride(50)
-                        .withSuppressConnectionHeader(true)
-                        .withKeepAliveOverride(true)
-                        .withCloseSocket(true)
+            new ConnectionOptions()
+                .withSuppressContentLengthHeader(true)
+                .withContentLengthHeaderOverride(50)
+                .withSuppressConnectionHeader(true)
+                .withKeepAliveOverride(true)
+                .withCloseSocket(true)
+                .withCloseSocketDelay(new Delay(SECONDS, 10))
         );
 
         // then
@@ -27,6 +30,7 @@ public class ConnectionOptionsDTOTest {
         assertThat(connectionOptions.getSuppressConnectionHeader(), is(true));
         assertThat(connectionOptions.getKeepAliveOverride(), is(true));
         assertThat(connectionOptions.getCloseSocket(), is(true));
+        assertThat(connectionOptions.getCloseSocketDelay(), is(new DelayDTO(new Delay(SECONDS, 10))));
     }
 
     @Test
@@ -38,6 +42,7 @@ public class ConnectionOptionsDTOTest {
         connectionOptions.setSuppressConnectionHeader(true);
         connectionOptions.setKeepAliveOverride(true);
         connectionOptions.setCloseSocket(true);
+        connectionOptions.setCloseSocketDelay(new DelayDTO(new Delay(SECONDS, 10)));
 
         // then
         assertThat(connectionOptions.getSuppressContentLengthHeader(), is(true));
@@ -45,6 +50,7 @@ public class ConnectionOptionsDTOTest {
         assertThat(connectionOptions.getSuppressConnectionHeader(), is(true));
         assertThat(connectionOptions.getKeepAliveOverride(), is(true));
         assertThat(connectionOptions.getCloseSocket(), is(true));
+        assertThat(connectionOptions.getCloseSocketDelay(), is(new DelayDTO(new Delay(SECONDS, 10))));
     }
 
     @Test
@@ -58,6 +64,7 @@ public class ConnectionOptionsDTOTest {
         assertThat(connectionOptions.getSuppressConnectionHeader(), nullValue());
         assertThat(connectionOptions.getKeepAliveOverride(), nullValue());
         assertThat(connectionOptions.getCloseSocket(), nullValue());
+        assertThat(connectionOptions.getCloseSocketDelay(), nullValue());
     }